+++ packages/default-components/src/shared/components/index.ts
@@ -1,0 +1,8 @@
+export { default as CustomInteractionFallback } from './CustomInteractionFallback.svelte';
+export { default as DrawingCanvas } from './DrawingCanvas.svelte';
+export { default as FileUpload } from './FileUpload.svelte';
+export { default as GraphicGapMatch } from './GraphicGapMatch.svelte';
+export { default as MatchDragDrop } from './MatchDragDrop.svelte';
+export { default as MathContent } from './MathContent.svelte';
+export { default as ModalFeedbackDisplay } from './ModalFeedbackDisplay.svelte';
+export { default as SortableList } from './SortableList.svelte';